// src/generated/schemas/updateOrderDetailStatusParams.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * GREGORI API
 * GREGORI 종합 쇼핑몰 REST API 명세서
 * OpenAPI spec version: 1.0.0
 */
import type { OrderDetailStatusUpdateDto } from './orderDetailStatusUpdateDto';

export type UpdateOrderDetailStatusParams = {
dto: OrderDetailStatusUpdateDto;
};