// src/generated/schemas/orderDetailStatusUpdateDtoStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * GREGORI API
 * GREGORI 종합 쇼핑몰 REST API 명세서
 * OpenAPI spec version: 1.0.0
 */

export type OrderDetailStatusUpdateDtoStatus = typeof OrderDetailStatusUpdateDtoStatus[keyof typeof OrderDetailStatusUpdateDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrderDetailStatusUpdateDtoStatus = {
  PAYMENT_CANCELED: 'PAYMENT_CANCELED',
  PAYMENT_COMPLETED: 'PAYMENT_COMPLETED',
  SHIPMENT_PREPARATION: 'SHIPMENT_PREPARATION',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
} as const;